backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database (SQLModel + Async PostgreSQL/SQLite)
sqlmodel>=0.0.22
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
greenlet>=3.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

# Google Gemini AI
google-genai>=1.0.0

# Validation and Settings
pydantic>=2.10.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.27.0
